agentic-samples/strands-builder-ui/requirements.txt
fastapi
uvicorn
sqlalchemy
strands-agents
strands-agents-tools
strands-agents[otel]
alembic
python-multipart
jinja2
pg8000
markdown
bleach
Pygments
markupsafe
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-rapidjson
gunicorn
uv
